add authentication

+++ app/api/post/delete/route.ts
@@ -1,9 +1,20 @@
-import { deletePost } from "@/lib/db/post";
+import { auth } from "@/lib/auth";
+import { deletePost, getPost } from "@/lib/db/post";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(req: NextRequest) {
+  const session = await auth();
+  if (!session) {
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+  }
+
+  const { id } = await req.json();
+  const targetPost = await getPost(id);
+  if (targetPost.author !== session.user?.name) {
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+  }
+  
   try {
-    const { id } = await req.json();
     deletePost(id);
     return NextResponse.json({message: 'Post deleted successfully'}, {status: 200});
   } catch (error: any) {
